+++ CHANGELOG.md
@@ -95,6 +95,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   parameter sets for an operation.
 - Typing of config vlaues for numpy parsed docstrings where type should be tuple
   or list
+- Model predict methods now use `SourcesContext.with_features`
 ### Removed
 - Monitor class and associated tests (unused)
 - DefinedFeature class in `dffml/feature/feature.py`